Добавить бронирование и комментарии

+++ src/main/java/ru/practicum/shareit/item/ItemStorage.java
@@ -4,9 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.shareit.item.model.Item;
 
+import java.math.BigInteger;
 import java.util.List;
 
-public interface ItemStorage extends JpaRepository<Item, Integer> {
+public interface ItemStorage extends JpaRepository<Item, BigInteger> {
 
     @Query(value = "select it from Item as it where it.owner.id = ?1")
     List<Item> findByUserId(Integer userId);
